requirements.txt
Flask
psycopg2-binary
flask_cors
load_dotenv